Implement q_delete_mid
This commit adds the q_delete_mid function, designed to delete the
middle node in a singly linked list that's being used as a queue,
as part of the lab0-c requirements.

The function first checks for edge cases: if the list is NULL or
empty, it returns false.  For non-empty lists, it initializes two
pointers: a fast pointer that starts two nodes ahead, and a slow
pointer that starts at the first data node.

The main logic resides in a while loop that continues as long as the
fast pointer can validly move forward two positions. Inside the loop,
the function checks two conditions, one is for odd-amount nodes,
another is for even-amount nodes. Depending on the different condition,
the function deletes the node. The slow pointer advances one position,
and the fast pointer advances two positions in each iteration.

Once the fast pointer reaches the end (or near the end) of the list,
the appropriate node is deleted, and the function returns true.

Change-Id: I44500e94a474fffc85401116aa75047397598456